Implement Menu Item Service and Repository

+++ RestaurantAPI.Domain/Entities/MenuItem.cs
@@ -4,6 +4,6 @@
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public int PriceCents { get; set; }
+        public decimal PriceCents { get; set; }
     }
 }
